% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{Barplot_FeatureSetEnrichmentAnalysis}
\alias{Barplot_FeatureSetEnrichmentAnalysis}
\title{Barplot of Feature Set Enrichment Analysis results}
\usage{
Barplot_FeatureSetEnrichmentAnalysis(fsea.out, alpha = 0.05)
}
\arguments{
\item{fsea.out}{output of \link{FeatureSetEnrichmentAnalysis} function}

\item{alpha}{FDR threshold fro calling enriched feature sets}
}
\value{
Barplot of number of enriched gene sets for each factors
}
\description{
Barplot of the Feature Set Enrichment Analyisis results
}
\details{
fill this
}
