% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProportionVariance.R
\name{CalculateProportionVariance}
\alias{CalculateProportionVariance}
\title{Calculate proportion of variance explained by each latent variable}
\usage{
CalculateProportionVariance(object, plot = F)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{plot}{generate a barplot}
}
\value{
a matrix with dim (M,K) with the proportion of residual variance explained by each factor. Barplot with the proportion of variance explained by the model for each view
}
\description{
Method to calculate the proportion of variance explained by each latent variable. \cr
There are several ways to do this:
Bayesian approach: we use the variance estimates from the variational distributions of the variables.
  var((E[W_k]*E[S_k])*E[Z_k])/var(Y)
Non-bayesian approach 1: to measure the proportion of variance explained by factor k we predict the data using only latent variable k
and we compute divide the predicted variance by the total variance (across samples and features).
   V_expl_k = var(Ypred_k)/var(Y)
Non-bayesian approach 2: one could argue that the non-bayesian approach 1 is not reliable because factors could have a lot of meaningless variance associated with it.
Therefore, better compute a goodness of fit such as the coefficient of determination:
  sum((Y[[m]]-Ypred_m[[m]])**2) / sum((Y[[m]]-mean(unlist(Y[[m]])))**2)
}
\details{
fill this
}

