% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProportionVariance.R
\name{CalculateVariance_Views}
\alias{CalculateVariance_Views}
\title{Calculate proportion of variance explained by each latent variable in each view}
\usage{
CalculateVariance_Views(object, views = "all", factors = "all",
  method = NULL)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{views}{Views to use, default is "all"}

\item{factors}{Latent variables or factores to use, default is "all"}

\item{method}{(ignore for now)}
}
\value{
a matrix with the proportion of variance explained by each factor (columns) in each view (rows)
}
\description{
Method to calculate the proportion of variance explained by each latent variable in each view. \cr
There are several ways to do this: \cr
Bayesian approach: we use the variance estimates from the variational distributions of the variables.\cr
  \eqn{var((E[W_k]*E[S_k])*E[Z_k])/var(Y)}
Non-bayesian approach 1: to measure the proportion of variance explained by factor k we predict the data using only latent variable k and we compute divide the predicted variance by the total variance (across samples and features): \cr
   \eqn{V_expl_k = var(Ypred_k)/var(Y)}
(default) Non-bayesian approach 2: one could argue that the non-bayesian approach 1 is not reliable because factors could have a lot of meaningless variance associated with it. Therefore, better compute a goodness of fit such as the coefficient of determination: \cr
  \eqn{sum((Y[[m]]-Ypred_m[[m]])**2) / sum((Y[[m]]-mean(unlist(Y[[m]])))**2)}
}
\details{
fill this
}
