% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{CLL_views}
\alias{CLL_views}
\title{CLL multi-omic data}
\format{A list containing filtered and pre-processed omic measurements for 211 CLL patients from Dietrich, Oles, Sellner et al including partially missing patients.
\itemize{
  \item{lincRNA}{normalized expression of lincRNA, p=2363}
  \item{mRNA}{normalized expression of mRNA, p=5000}
  \item{miRNA}{normalized expression of miRNA, p=274}
  \item{meth}{methylation M-values of top variable sites excluding sex chromosomes, p=4248}
  \item{viab}{drug response data for 62 drugs in 5 concentrations, p=310}
  \item{mut}{somatic mutations, IGHV status and copy number aberrations, p=69}
  \item{covariates}{diganosis and sex, p=2}
}}
\usage{
data(CLL_views)
}
\description{
A list containing differen omic measurements in CLL patient samples
}
