% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.R
\name{compare_models}
\alias{compare_models}
\title{Compare different trained MOFA object in terms of the final value of the ELBO statistics and number of inferred factors}
\usage{
compare_models(models, show_modelnames = FALSE)
}
\arguments{
\item{models}{a list containing \code{\link{MOFAmodel}} objects.}
}
\description{
Different objects of \code{\link{MOFAmodel}} are compared in terms of the final value of the ELBO statistics. 
For model selection the model with the highest ELBO value is selected.
}
