% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMOFAobject.R
\name{createMOFAobject}
\alias{createMOFAobject}
\title{Initialize a \code{\link{MOFAmodel}} object with multi-omic input data}
\usage{
createMOFAobject(mae, minViews = 1)
}
\arguments{
\item{mae}{A MultiAssayExperiment containing the input data for MOFA}

\item{minViews}{Number of views a sample has to be observed to be included (default 1, i.e. including all samples)}
}
\value{
an untrained MOFA model object
}
\description{
Method to intilize the InputData slot of the MOFA object starting from a MutliAssayExperiment object
}
\details{
fill this
}
