% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectPassengers.R
\name{detectPassengers}
\alias{detectPassengers}
\title{detectPassengers:}
\usage{
detectPassengers(object, views = "all", factors = "all",
  r2_threshold = 0.03)
}
\arguments{
\item{object}{a MOFA model}

\item{views}{all}

\item{factors}{all}

\item{r2_threshold}{minimal level of rel. R2 for a factor to be considered active one a view}
}
\value{
fill this
}
\details{
fill this
}
