% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFactors.R
\name{FactorBeeswarmPlot}
\alias{FactorBeeswarmPlot}
\title{Visualize beeswarm plot of one latent variable}
\usage{
FactorBeeswarmPlot(object, factors, color_by = NULL, name_color = "",
  showMissing = TRUE)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{factors}{factors to plot}

\item{color_by}{specifies groups or values used to color points. This can be either a character giving the name of a feature or covariate or a vector of same length as number of samples specifying a group or value for each sample.}

\item{name_color}{name for color legend (usually only used if color_by is not a character itself)}

\item{showMissing}{boolean, if false, removes sample for which shape_by or color_by is missing}
}
\value{
ggplot object containing the beeswarm plots
}
\description{
fill this
}
\details{
asd
}
\references{
fill this
}
