% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFactors.R
\name{FactorHistPlot}
\alias{FactorHistPlot}
\title{Visualize histogram of one latent variable}
\usage{
FactorHistPlot(object, factor, groups = NULL, name_groups = "",
  alpha = 0.6, binwidth = NULL, showMissing = FALSE)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{factor}{one factors to plot}

\item{groups}{specifies groups or values used to split histogram into groups and color. This can be either a character giving the name of a feature or covariate or a vector of same length as number of samples specifying a group or value for each sample.}

\item{name_groups}{name for groups (usually only used if groups is not a character itself)}

\item{alpha}{transparency parameter}

\item{binwidth}{binwidth for histogram}

\item{showMissing}{boolean, if false, removes sample for which shape_by or color_by is missing}
}
\value{
fill this
}
\description{
fill this
}
\details{
asd
}
\references{
fill this
}
