% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{factorNames}
\alias{factorNames}
\alias{factorNames}
\alias{factorNames<-}
\alias{factorNames,MOFAmodel-method}
\alias{factorNames<-,MOFAmodel,vector-method}
\title{factorNames: set and retrieve factor names}
\usage{
factorNames(object)

factorNames(object) <- value

\S4method{factorNames}{MOFAmodel}(object)

\S4method{factorNames}{MOFAmodel,vector}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{value}{a character vector of factor names}
}
\value{
character vector with the features names
}
