% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewVsFactorPlot.R, R/corPlot.R
\name{FactorsCorPlot}
\alias{FactorsCorPlot}
\alias{ViewFactorPlot}
\title{View versus Factor plot}
\usage{
ViewFactorPlot(model, ...)

FactorsCorPlot(object, method = "pearson", ...)
}
\arguments{
\item{...}{arguments passed to \code{corrplot}}

\item{object}{a \code{\link{MOFAmodel}} object.}

\item{method}{a character string indicating which correlation coefficient is to be computed: pearson (default), kendall, or spearman.}

\item{object}{a \code{\link{MOFAmodel}} object.}

\item{method}{a character string indicating which correlation coefficient is to be computed: pearson (default), kendall, or spearman.}

\item{...}{arguments passed to \code{corrplot}}
}
\value{
fill this

fill this
}
\description{
fill this

fill this
}
\details{
asd

asd
}

