% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPlot.R
\name{FeaturesCorPlot}
\alias{FeaturesCorPlot}
\title{Visualize correlation matrix between the features}
\usage{
FeaturesCorPlot(object, view, method = "pearson", regress_factors = NULL,
  top = 500, ...)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
fill this
}
\description{
fill this
}
\details{
asd
}
\references{
fill this
}
