% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMOFA.R
\name{getDefaultDatasOpts}
\alias{getDefaultDatasOpts}
\alias{getDefaultDataOpts}
\title{getDefaultDataOpts: Get default data options}
\usage{
getDefaultDataOpts()
}
\value{
list with data options
}
\description{
Function to obtain default data options
}
