% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMOFA.R
\name{getDefaultModelOpts}
\alias{getDefaultModelOpts}
\title{getDefaultModelOpts: Get default model options}
\usage{
getDefaultModelOpts(object)
}
\arguments{
\item{object}{untrained MOFA object to get model options for}
}
\value{
list with default model options
}
\description{
Function to obtain default model options
}
