% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMOFA.R
\name{getDefaultTrainOpts}
\alias{getDefaultTrainOpts}
\title{Get default training options}
\usage{
getDefaultTrainOpts()
}
\value{
list with default training options
}
\description{
Function to obtain default training options
}
\details{
The training options are the following: \cr
maxiter: Maximum number of iterations. \cr
tolerance: Convergence threshold based on the change in Evidence Lower Bound, we recommend this be around 0.01. \cr
DropFactorThreshold: Threshold on fraction of variance explained to drop a factor. That is, factors explaining less than 'DropFactorThreshold' fraction of variance (in all views) will be dropped. \cr
}
