% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getDimensions}
\alias{getDimensions}
\title{wraper to extract dimensionalities from the MOFAmodel.}
\usage{
getDimensions(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\description{
K indicates the number of factors, D indicates the number of features, N indicates the number of samples and M indicates the number of views.
}
