% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getELBO}
\alias{getELBO}
\title{getELBO}
\usage{
getELBO(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
value of the ELBO statistic at end of training
}
\description{
Extract the value of the ELBO statistics after model training.
 This can be useful for model selection.
}
\examples{
# load a trained MOFAmodel object
filepath <- system.file("extdata", "scMT_model.hdf5", package = "MOFAdata")
MOFAobject <- loadModel(filepath)
# get ELBO statistic
getELBO(MOFAobject)
}
