% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getExpectations}
\alias{getExpectations}
\title{Fetch expectations from the model}
\usage{
getExpectations(object, variable, as.data.frame = FALSE)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{variable}{variable name, 'Z' for factors, 'SW' for weights, 'Tau' for noise, 'Y' for pseudodata, 'Theta' for feature-wise spike-and-slab sparsity, 'AlphaW' for view and factor-wise ARD sparsity}

\item{as.data.frame}{boolean indicating whether to output the result as a long data frame, default is FALSE.}
}
\value{
the output varies depending on the variable of interest: \cr
Z: a matrix with dimensions (samples,factors). If as.data.frame is TRUE, a long-formatted data frame with columns (sample,factor,value). \cr
SW: a list of length (views) where each element is a matrix with dimensions (features,factors). If as.data.frame is TRUE, a long-formatted data frame with columns (view,feature,factor,value). \cr
Y: a list of length (views) where each element is a matrix with dimensions (features,samples). If as.data.frame is TRUE, a long-formatted data frame with columns (view,feature,sample,value). \cr
Theta: TO-FILL \cr
Tau: TO-FILL
}
\description{
this method extracts the expectations from the (variational) posterior distributions of a trained MOFAmodel.
}
\details{
Technical note: MOFA is a Bayesian model where each variable has a prior distribution and a posterior distribution. 
In particular, to gain speed we used the variational inference framework so true posterior distributions are replaced by approximated variational distributions.
The priors and variational distributions of each variable are extensively described in the supplementary methods of the paper.
}
