% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGoodnessOfFit.R
\name{getGoodnessOfFit}
\alias{getGoodnessOfFit}
\title{Get a goodness of fit measure for the model in each view and for each LF}
\usage{
getGoodnessOfFit(object, views = "all", factors = "all", plotit = T,
  perFeature = F, interceptLF = F, allAsGaussian = T)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{views}{Views to use, default is "all"}

\item{factors}{Latent variables or factores to use, default is "all"}

\item{plotit}{boolean, wether to produce a plot (default true)}

\item{perFeature}{boolean, whether to calculate in addition variance explained per feature (and factor) (default FALSE)}
}
\value{
a list containing list of R2 for all gaussian views and list of BS for all binary views
}
\description{
Method to calculate a a goodness of fit measure for the model in each view and for each LF
For gaussian views the coefficient of determination is used as a measure and 
is calculated for the whole model as well as for individual factors on each view and feature.
For binary views the Brier score is used  as a measure and 
is calculated for the whole model as well as for individual factors on each view and feature.
Other families are not yet implemented.
The resulting measures per factor and overall are plotted as a heatmap and barplot by default.
NOTE: The two measure do not measure the same thing and should not be compared!
}
\details{
fill this
}
