% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getImputedData}
\alias{getImputedData}
\title{Fetch the imputed data}
\usage{
getImputedData(object, views = "all", features = "all", as.data.frame = F)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{views}{character vector with the view name(s), or numeric vector with the view index(es), default is "all".}

\item{features}{list of character vectors with the feature names or list of numeric vectors with the feature indices, default is "all"}

\item{as.data.frame}{boolean indicating whether to return a long-formatted data frame instead of a list of matrices, default is FALSE.}
}
\value{
by default returns a list where each element is a matrix with dimensionality (D,N) where D is the number of features in this view and N is the number of samples.
Alternatively, if as.data.frame is TRUE, returns a long-formatted data frame with columns (view,feature,sample,value).
}
\description{
collect the imputed data. It requires the previous use of the \code{\link{imputeMissing}} method.
}
