% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getParameters}
\alias{getParameters}
\title{wraper to fetch particular parameters from the model}
\usage{
getParameters(object, variable, parameter, as.data.frame = F)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{variable:}{variable name ('AlphaW')}

\item{parameter:}{parameter name ('a')}

\item{as.data.frame:}{output the result as a long data frame?}
}
\description{
to-fill
}
