% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getTrainData}
\alias{getTrainData}
\title{wraper to fetch training data from the model}
\usage{
getTrainData(object, views = "all", as.data.frame = F)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{views:}{views (default is "all")}

\item{as.data.frame:}{output the result as a long data frame?}
}
\description{
to-fill
}
