% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getWeights}
\alias{getWeights}
\title{wraper to extract the weights from the model}
\usage{
getWeights(object, views = "all", factors = "all", as.data.frame = F)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{views}{vector containing the names of views (character) or index of views (numeric) to be fetched.}

\item{factors}{vector with the factors indices (numeric) or factor names (character) to fetch.}

\item{as.data.frame:}{boolean indicating whether to return factors as a long data frame with columns (view,feature,factor,value), default is FALSE.}
}
\value{
by default returns a list of length M, where M is the number of views, of matrices with dimensionality (Dm,K) where Dm is number of features in view m and k is number of factors. 
Alternatively, with as.data.frame=TRUE, it returns a data frame with columns (view,feature,factor,value)
}
\description{
to-fill
}
