% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEA}
\alias{GSEA}
\title{Gene Set Enrichment Analysis}
\usage{
GSEA(model, view, factor.indexes, gene.sets, local.statistic = "loading",
  transformation = "none", global.statistic = "mean.diff",
  statistical.test = "cor.adj.parametric", nperm = 1, min_size = 10)
}
\arguments{
\item{model}{a \code{\link{MOFAmodel}} object.}

\item{view}{name (character) or index (integer) of the view of interest}

\item{factor.indexes}{numeric vector with the indices of the latent variables for which enrichment should be computed.}

\item{gene.sets}{Data structure that holds gene set membership information. Must be either a binary membership matrix (rows are gene sets and columns are genes) or a list of gene set member indexes.}

\item{local.statistic}{The gene-level statistic used to quantify the association between each genomic variable and each PC. Must be one of the following: loading (default), cor, z.}

\item{transformation}{Optional transformation to apply to the gene-level statistics. Must be one of the following "none" (default) or "abs.value".}

\item{global.statistic}{The gene set statisic computed from the gene-level statistics. Must be one of the following: "mean_diff" (default) or "rank.sum".}

\item{statistical.test}{The statistical test used to compute the significance of the gene set statistics under a competitive null hypothesis.
Must be one of the following: "parametric", "cor.adj.parametric" (default), "permutation".}

\item{nperm}{Number of permutations to perform. Only relevant if statistical.test is "permutation".}
}
\value{
a list with two matrices, one for statistics and one for pvalues. The rows are the different gene sets and the columns are the latent variables.
}
\description{
Method to perform gene ontology enrichment for a gene sets in a given view  \cr
Here we use the PCGSE package.
}
\details{
fill this
}
