% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{Heatmap_FeatureSetEnrichmentAnalysis}
\alias{Heatmap_FeatureSetEnrichmentAnalysis}
\title{Heatmap of Feature Set Enrichment Analysis results}
\usage{
Heatmap_FeatureSetEnrichmentAnalysis(fsea.out, threshold = 0.05, log = T,
  adjust = TRUE, ...)
}
\arguments{
\item{fsea.out}{output of \link{FeatureSetEnrichmentAnalysis} function}

\item{threshold}{p.value threshold to filter out feature sets. If a feature set has a p.value lower than 'threshold'}

\item{adjust}{use multiple testing correction}

\item{...}{Parameters to be passed to pheatmap function}
}
\value{
vector of factors being enriched for at least one feautre set at the threshold specified
}
\description{
Heatmap of the Feature Set Enrichment Analyisis results
}
\details{
fill this
}
