% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{ImputedData}
\alias{ImputedData}
\alias{ImputedData}
\alias{ImputedData<-}
\alias{ImputedData,MOFAmodel-method}
\title{ImputedData: set and retrieve imputed data}
\usage{
ImputedData(object)

.ImputedData(object) <- value

\S4method{ImputedData}{MOFAmodel}(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
list of numeric matrices that contain the imputed data
}
\description{
Function to set and retrieve imputed data.
}
\examples{
# load a trained MOFAmodel object
filepath <- system.file("extdata", "CLL_model.hdf5", package = "MOFAtools")
MOFAobject <- loadModel(filepath)
# impute missing values
MOFAobject <- imputeMissing(MOFAobject)
# get imputations for a single view
imputedDrugs <-ImputedData(MOFAobject)$Drugs
head(imputedDrugs)
}
