% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{InputData}
\alias{InputData}
\alias{InputData,MOFAmodel-method}
\title{InputData: set and retrieve training data}
\usage{
InputData(object)

\S4method{InputData}{MOFAmodel}(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
a MultiAssayExperiment containign all input data and metadata
}
