% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{LinePlot_FeatureSetEnrichmentAnalysis}
\alias{LinePlot_FeatureSetEnrichmentAnalysis}
\title{Line plot of Feature Set Enrichment Analysis results}
\usage{
LinePlot_FeatureSetEnrichmentAnalysis(p.values, view, factor, threshold = 0.1,
  max.pathways = 25, ...)
}
\arguments{
\item{p.values}{output of \link{FeatureSetEnrichmentAnalysis} function. A data frame of p.values where rows are feature sets and columns are latent variables}

\item{threshold}{p.value threshold to filter out feature sets}

\item{max.pathways}{maximum number of enriched pathways to display}
}
\value{
nothing
}
\description{
Line plot of the Feature Set Enrichment Analyisis results for a specific latent variable
}
\details{
fill this
}
