% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModel.R
\name{loadModel}
\alias{loadModel}
\title{LoadModel: loading a MOFA model}
\usage{
loadModel(file)
}
\arguments{
\item{file}{an hd5f object generated by the MOFA python package.}
}
\value{
a \code{\link{MOFAmodel}} object.
}
\description{
Method to load a trained MOFA model from a hdf5 file. \cr
Models have to be trained and saved using the corresponding Python package.
}
\details{
fill this
}
