% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{Parameters}
\alias{Parameters}
\alias{Parameters,MOFAmodel-method}
\title{Parameters: set and retrieve parameters}
\usage{
Parameters(object)

\S4method{Parameters}{MOFAmodel}(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
value list of Parameters
}
\author{
Ricard Argelaguet
}
