% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFactors.R
\name{PlotCorFactors}
\alias{PlotCorFactors}
\title{Plot the correlation matrix between the latent factors}
\usage{
PlotCorFactors(object, method = "pearson", ...)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{method}{a character string indicating which correlation coefficient is to be computed: pearson (default), kendall, or spearman.}

\item{...}{arguments passed to \code{corrplot}}
}
\value{
symmetric matrix with the correlation coefficient between every pair of factors
}
\details{
this method plots the correlation matrix between the latent factors. \cr 
Factors are encouraged to be uncorrelated due to the assumptions of the model. However, it is not a hard constraint such as in Principal Component Analysis and correlations between factors can happen, particularly with large number factors.
Generally, correlated factors are redundant and should be avoided, as they make interpretation harder. Therefore, if you have too many correlated factors we suggest you run the model again reducing the number of factors.
}
