% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFactors.R
\name{plotFactorBeeswarm}
\alias{plotFactorBeeswarm}
\title{Visualize beeswarm plot of one latent variable}
\usage{
plotFactorBeeswarm(object, factors, color_by = NULL, name_color = "",
  showMissing = FALSE)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{factors}{factors to plot}

\item{color_by}{specifies groups or values used to color points. This can be either a character giving the name of a feature or covariate or a vector of same length as number of samples specifying a group or value for each sample.}

\item{name_color}{name for color legend (usually only used if color_by is not a character itself)}

\item{showMissing}{boolean, if false, removes sample for which shape_by or color_by is missing}
}
\value{
ggplot object
}
\description{
fill this
}
\details{
One of the main steps for the annotation of factors is to visualise and group/color them using known covariates or phenotypic data.
This method generates a beeswarm plot of the sample values in a given latent factor. Note that, similar to Principal Component Analysis, the factor values should be interpreted in a relative manner.
Similar functions are \code{\link{plotFactorsScatter}} for doing scatter plots and \code{\link{plotFactorHist}} for doing histogram plots
}
