% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotTilesData}
\alias{plotTilesData}
\title{plotTilesData: tile plot of data availability for each sample}
\usage{
plotTilesData(object, cols = NULL)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{cols}{a vector of length M (number of views in object) specifying colors to use}
}
\description{
Function to do a tile plot showing the availability of data per view and sample
}
\details{
This function is helpful to get an overview of the training data used for MOFA. 
For each sample it shows in which view data is available.
}
