% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWeights.R
\name{plotTopWeights}
\alias{plotTopWeights}
\title{Plot top weights}
\usage{
plotTopWeights(object, view, factor, nfeatures = 10, abs = TRUE,
  scale = TRUE, sign = "both")
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{view}{character vector with the voiew name, or numeric vector with the index of the view to use.}

\item{factor}{character vector with the factor name, or numeric vector with the index of the factor to use.}

\item{nfeatures}{number of top features to display.}

\item{abs}{logical indicating whether to use the absolute value of the weights.}

\item{scale}{logical indicating whether to scale all loadings from 0 to 1.}

\item{sign}{can be 'positive', 'negative' or 'both' to show only positive, negative or all weigths, respectively.}
}
\description{
An important step to annotate factors is to visualise the corresponding feature loadings. \cr
This function displays the top features with highest loading. 
In contrast, the function \code{\link{plotTopWeights}} plots all loadings for a given latent factor and view.
}
\details{
The weights of the features within a view are relative andthey should not be interpreted in an absolute scale.
Therefore, for interpretability purposes we always recommend to scale the weights with \code{scale=TRUE}.
}
