% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWeights.R
\name{plotWeights}
\alias{plotWeights}
\title{plotWeights: visualise the distribution of loadings for a certain factor in a given view}
\usage{
plotWeights(object, view, factor, nfeatures = 10, abs = FALSE,
  manual = NULL, color_manual = NULL, scale = TRUE)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{view}{name of view from which to get the corresponding weights}

\item{factor}{name of the factor}

\item{nfeatures}{number of features to label}

\item{abs}{take absolute value of the weights?}

\item{manual}{features to be manually labelled. A list of character vectors.}

\item{color_manual}{a character vector with colors, one for each element of 'manual'}

\item{scale}{boolean, scale all loadings from 0 to 1?}
}
\description{
The first step to annotate factors is to visualise the corresponding feature loadings. \cr
This function shows all loadings for a given latent factor and view. . \cr
In contrast, the function \code{showTopWeights} zooms and displays the features with highest loading.
}
\details{
the weights of the features within a view are relative to each other, they should not be interpreted in an absolute scale
}
