% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWeights.R
\name{plotWeightsHeatmap}
\alias{plotWeightsHeatmap}
\title{plotWeightsHeatmap: show heatmap of the factor loadings in a specific view}
\usage{
plotWeightsHeatmap(object, view, features = "all", factors = "all",
  threshold = 0, color = NULL, breaks = NULL, ...)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{view}{view name}

\item{features}{name of the features to include (default: "all")}

\item{factors}{name of the factors to include (default: "all")}

\item{threshold}{threshold on absolute weight values, features/factors that are below this threshold in all factors/features are removed}

\item{color}{to fill}

\item{breaks}{to fill}

\item{...}{further arguments that can be passed to pheatmap}
}
\value{
fill this
}
\description{
Function to visualize loadings for a given set of factors in a given view. This is useful to see the overall pattern of the weights but it is not very useful to characterise the factors.
 If you want to individually look at the top weights for a given factor we recommend the function showAllWeights and showTopWeights
}
\details{
fill this
}
