% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMOFA.R
\name{prepareMOFA}
\alias{prepareMOFA}
\title{prepareMOFA}
\usage{
prepareMOFA(object, DirOptions, DataOptions = NULL, ModelOptions = NULL,
  TrainOptions = NULL)
}
\arguments{
\item{object}{an untrained \code{\link{MOFAmodel}}}

\item{DirOptions}{list with I/O options, it must contain a 'dataDir' element where temporary text files will be stored 
and an 'outFile' with hdf5 extension where the final model will be stored.}

\item{DataOptions}{list of DataOptions (see \code{\link{getDefaultDataOpts}} details). 
If NULL, default data options are used.}

\item{ModelOptions}{list of ModelOptions (see \code{\link{getDefaultModelOpts}} for details). 
If NULL, default model options are used.}

\item{TrainOptions}{list of TrainOptions (see \code{\link{getDefaultTrainOpts}} for details). 
If NULL, default training options are used.}
}
\value{
a untrained \code{\link{MOFAmodel}} with specified DataOpts, ModelOpts and TrainOpts
}
\description{
Function to prepare a \code{\link{MOFAmodel}} object for training. It requires defining data, model and training options.
}
