% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMOFA.R
\name{prepareMOFAInputFiles}
\alias{prepareMOFAInputFiles}
\title{prepareMOFAInputFiles: Write omics .txt files for input to Python}
\usage{
prepareMOFAInputFiles(object, dir)
}
\arguments{
\item{object}{an untrained MOFA object}

\item{dir}{directory to store .txt in}
}
\value{

}
\description{
Function to produce .txt files that are used for python as input. These files are all stored in the specified directory.
}
\details{
fill this
}
