% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMOFA.R
\name{prepareMOFARunFile}
\alias{prepareMOFARunFile}
\alias{runMOFA}
\title{prepareMOFARunFile: Write .sh files to run MOFA in Python}
\usage{
runMOFA(object, DirOptions)
}
\arguments{
\item{object}{an untrained MOFA object}

\item{dir}{directory to store .txt in}

\item{outFile}{name of output file from Python MOFA}

\item{k}{number of latent factors to start with (default = 10)}

\item{MOFAdir}{directory of the MOFA Pyhton package installation}
}
\value{

}
\description{
Function to produce .sh files to run MOFA in Python from the command line with specified options.
}
\details{
fill this
}
