% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{qualityControl}
\alias{qualityControl}
\title{Quality control of a MOFA model}
\usage{
qualityControl(object, verbose = F)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
a \code{\link{MOFAmodel}} model.
}
\description{
(TO-FILL) Method to do quality control on a MOFA model. \cr
}
