% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames}
\alias{sampleNames<-}
\alias{sampleNames,MOFAmodel-method}
\alias{sampleNames<-,MOFAmodel,vector-method}
\title{sampleNames: set and retrieve sample names}
\usage{
sampleNames(object)

sampleNames(object) <- value

\S4method{sampleNames}{MOFAmodel}(object)

\S4method{sampleNames}{MOFAmodel,vector}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{value}{a character vector of sample names}
}
\value{
character vector with the sample names
}
