% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.R
\name{selectModel}
\alias{selectModel}
\title{Select a model from a list of trained MOFAobjects based on the best ELBO value}
\usage{
selectModel(models, plotit = TRUE)
}
\arguments{
\item{models}{a list containing \code{\link{MOFAmodel}} objects.}

\item{plotit}{logical indicating whether to plot the model selection.}
}
\description{
Different objects of \code{\link{MOFAmodel}} are compared in terms of the final value of the ELBO statistics 
and the model with the highest ELBO value is selected.
}
