% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showWeights.R
\name{showAllWeights}
\alias{showAllWeights}
\title{showAllWeights: visualise the distribution of loadings for a certain factor in a given view}
\usage{
showAllWeights(object, view, factor, nfeatures = "all", abs = FALSE,
  threshold = NULL, manual = NULL, color_manual = NULL, main = NULL)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{view}{name of view from which to get the corresponding weights}

\item{factor}{name of the factor}

\item{nfeatures}{number of features to include (by default all, otherwise only those with highest absolute values are included)}

\item{abs}{take absolute value of the weights?}

\item{threshold}{threshold on the absolute value beyond which weigths are labeled}

\item{manual}{features to be manually labelled. A list of character vectors}

\item{color_manual}{a character vector with colors}

\item{main}{plot title}
}
\description{
The first step to annotate factors is to visualise the corresponding feature loadings. \cr
This function shows all loadings for a given latent factor and view. . \cr
In contrast, the function \code{showTopWeights} zooms and displays the features with highest loading.
}
\details{
fill this
}
