% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSparsity.R
\name{showSparsity}
\alias{showSparsity}
\title{showSparsity: show the sparsity of each factor in each view}
\usage{
showSparsity(model, views = "all", factors = "all", main = NULL)
}
\arguments{
\item{model}{a MOFA model}

\item{views}{character vector of view names (default: "all")}

\item{factors}{character vector of factor names (default: "all")}
}
\value{
fill this
}
\description{
Function to visualize the sparsity of a given factor(s) for a given view(s)
}
\details{
fill this
}
