% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showWeights.R
\name{showTopWeights}
\alias{showTopWeights}
\title{showTopWeights: visualise the top weights for a certain factor in a given view}
\usage{
showTopWeights(object, view, factor, nfeatures = 5, manual_features = NULL,
  sign = "positive", abs = TRUE, scale = T)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{view}{character vector with the view name or numeric vector with the index of the view.}

\item{factor}{character vector with the factor name or numeric vector with the index of the factor.}

\item{nfeatures}{number of features}

\item{manual_features}{names of features to be included into the plot. If null (default) take top features as specified in nfeatures.}

\item{sign}{can be 'positive', 'negative' or 'both' to show only positive, negative or all weigths, respectively}

\item{abs}{take absolute value of the weights?}

\item{scale}{scales all loading by absolute value of maximal loading}
}
\description{
The first step to annotate factors is to visualise the corresponding feature loadings. \cr
This function zooms and displays the features with highest loading in a given latent factor and a given view. \cr
In contrast, the function \code{showAllWeights} displays the entire distribution of weights.
}
