% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showWeights.R
\name{showWeights}
\alias{showWeights}
\title{showWeights: plot weights for a certain factor and view}
\usage{
showWeights(model, view, factor, ntop = 5, ntail = 5, manual = NULL)
}
\arguments{
\item{model}{a fitted MOFA model}

\item{view}{name of view from which to get the corresponding weights}

\item{factor}{factor to plot weights for}

\item{ntop}{number of highest positive weights to label}

\item{ntail}{number of lowest negative weights to label}
}
\value{
a dataframe with of d_m rows (feautres) with weights per feature on the sepcified factor and view
}
\description{
Function to visualize weights in a dotplot highlighting top positive and neagtive featurs
}
\details{
fill this
}
