% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetFactors.R
\name{subsetFactors}
\alias{subsetFactors}
\title{subsetFactors: subset factors}
\usage{
subsetFactors(object, factors)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{factors}{numeric or character vector with the factors indices (numeric) or factor names (character) to subset or sort}
}
\value{
nothing
}
\description{
Method to subset factors
}
\details{
fill this
}
