% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetFactors.R
\name{subsetSamples}
\alias{subsetSamples}
\title{subsetSamples: subset samples}
\usage{
subsetSamples(object, samples)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{samples}{numeric or character vector with the sample indices (numeric) or sample names (character) to subset (or sort)}
}
\description{
Method to subset (or sort) samples
}
