% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingStatistics.R
\name{trainCurve}
\alias{trainCurve}
\title{Training Statistic}
\usage{
trainCurve(object, statistic, title = "", titlesize = 16, xlabel = "Time",
  ylabel = "", linesize = 1.2, xlim_down = NA, xlim_up = NA,
  ylim_down = NA, ylim_up = NA)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{statistic}{fill this}
}
\description{
to-fill
}
