% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingStatistics.R
\name{trainCurveELBO}
\alias{trainCurveELBO}
\title{Training curve for Evidence Lower Bound (ELBO)}
\usage{
trainCurveELBO(object, log = F, title = "", titlesize = 16,
  xlabel = "Iteration", ylabel = NULL, linesize = 1.2, xlim_down = NULL,
  xlim_up = NULL)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{statistic}{fill this}
}
\description{
to-fill
}
