% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingStatistics.R
\name{trainCurveFactors}
\alias{trainCurveFactors}
\title{Training Curve for number of active factors}
\usage{
trainCurveFactors(object, title = "", titlesize = 16,
  xlabel = "Iteration", ylabel = NULL, linesize = 1.2, xlim_down = NULL,
  xlim_up = NULL)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\description{
to-fill
}
