% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{TrainData}
\alias{TrainData}
\alias{TrainData}
\alias{TrainData<-}
\alias{TrainData,MOFAmodel-method}
\title{TrainData: set and retrieve training data}
\usage{
TrainData(object)

.TrainData(object) <- value

\S4method{TrainData}{MOFAmodel}(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
list of numeric matrices that contain the training data
}
\description{
Function to set and retrieve training data.
}
\examples{
data("scMT_data")
MOFAobject <- createMOFAobject(scMT_data)
TrainData(MOFAobject)

data("CLL_data")
MOFAobject <- createMOFAobject(CLL_data)
TrainData(MOFAobject)
}
