% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{TrainOptions}
\alias{TrainOptions}
\alias{TrainOptions}
\alias{TrainOptions<-}
\alias{TrainOptions,MOFAmodel-method}
\title{TrainOptions: set and retrieve training opts}
\usage{
TrainOptions(object)

.TrainOptions(object) <- value

\S4method{TrainOptions}{MOFAmodel}(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
list of training options
}
\description{
Function to set and retrieve training options.
}
\examples{
# load existing object
filepath <- system.file("extdata", "scMT_model.hdf5", package = "MOFAtools")
MOFAobject <- loadModel(filepath)
# get TrainOptions
TrainOptions(MOFAobject)
}
