% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{TrainStats}
\alias{TrainStats}
\alias{TrainStats}
\alias{TrainStats<-}
\alias{TrainStats,MOFAmodel-method}
\title{TrainStats: set and retrieve training statistics}
\usage{
TrainStats(object)

.TrainStats(object) <- value

\S4method{TrainStats}{MOFAmodel}(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
list of training statistics
}
\description{
Function to set and retrieve training statistics
}
\examples{
# load existing object
filepath <- system.file("extdata", "scMT_model.hdf5", package = "MOFAtools")
MOFAobject <- loadModel(filepath)
# get training statistics
trainStats <- TrainStats(MOFAobject)
}
