% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewVsFactorPlot.R
\name{ViewFactorPlot}
\alias{ViewFactorPlot}
\title{View versus Factor plot}
\usage{
ViewFactorPlot(model, ...)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
fill this
}
\description{
fill this
}
\details{
asd
}
