% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{beeswarmPlot}
\alias{beeswarmPlot}
\title{Visualize beeswarm plot of one latent variable}
\usage{
beeswarmPlot(object, factor, label = NULL, color = NULL, color_name = "",
  legend = F)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{id}{latent variable}
}
\value{
fill this
}
\description{
fill this
}
\details{
asd
}
\references{
fill this
}
