% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateVarianceExplained.R
\name{calculateVarianceExplained}
\alias{calculateVarianceExplained}
\title{Calculate variance explained by the model}
\usage{
calculateVarianceExplained(object, views = "all", factors = "all",
  plotit = T, perFeature = F, orderFactorsbyR2 = F, showtotalR2 = T,
  showVarComp = F)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{views}{Views to use, default is "all"}

\item{factors}{Latent factores to use, default is "all"}

\item{plotit}{boolean, wether to produce a plot (default True)}

\item{perFeature}{boolean, whether to calculate in addition variance explained per feature (and factor) (default FALSE)}

\item{orderFactorsbyR2}{order factors according to sum of variance explained across views (default TRUE)}

\item{showtotalR2}{Calculate R2 with respect to the total variance (TRUE) or the residual variance (FALSE?}

\item{showVarComp}{Calculate R2 per view using all factors (default TRUE)}
}
\value{
a list with matrices with the amount of variation explained per factor and view, and optionally total variance explained per view and variance explained by each feature alone
}
\description{
Method to calculate variance explained by the MOFA model for each view and latent factor, and optionally also for each feature.
As a measure of variance explained we adopt the coefficient of determination (R2).
For non-gaussian views the calculations are based on the normally-distributed pseudo-data (for more information on the non-gaussian model see Seeger & Bouchard, 2012).
}
\details{
fill this
}
