% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateVarianceExplained.R
\name{calculateVarianceExplained}
\alias{calculateVarianceExplained}
\title{Calculate variance explained in a MOFA model for each view and latent factor}
\usage{
calculateVarianceExplained(object, views = "all", factors = "all",
  plotit = T, perFeature = F, InterceptLFasNull = T,
  orderFactorsbyR2 = T, showtotalR2 = T)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{views}{Views to use, default is "all"}

\item{factors}{Latent variables or factores to use, default is "all"}

\item{plotit}{boolean, wether to produce a plot (default true)}

\item{perFeature}{boolean, whether to calculate in addition variance explained per feature (and factor) (default FALSE)}
}
\value{
a list containing list of R2 for all gaussian views and list of BS for all binary views
}
\description{
Method to calculate variance explained in a MOFA model for each view and latent factor.
As a measure of variance explained the coefficient of determination, given by 1 - SS_res/SS_total.
For non-gaussian views the calculations are basedon the gaussian pseudo-data. 
The resulting measures per factor and overall are plotted as a heatmap and barplot by default.
}
\details{
fill this
}
