% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateVarianceExplained.R
\name{calculateVarianceExplained}
\alias{calculateVarianceExplained}
\title{Calculate variance explained by the model}
\usage{
calculateVarianceExplained(object, views = "all", factors = "all",
  perFeature = F, perView = F, totalVar = T, plotit = T)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{views}{Views to use, default is "all"}

\item{factors}{Latent factores to use, default is "all"}

\item{perFeature}{boolean, whether to calculate in addition the variance explained (R2) per feature (default FALSE)}

\item{perView}{boolean, whether to calculate in addition the variance explained (R2) per view, using all factors (default TRUE)}

\item{totalVar}{calculate variance explained (R2) with respect to the total variance (TRUE) or the residual variance (FALSE?}

\item{plotit}{boolean, wether to produce a plot (default True)}
}
\value{
a list with matrices with the amount of variation explained per factor and view, and optionally total variance explained per view and variance explained by each feature alone
}
\description{
Method to calculate variance explained by the MOFA model for each view and latent factor, and optionally also for each feature.
As a measure of variance explained we adopt the coefficient of determination (R2).
For non-gaussian views the calculations are based on the normally-distributed pseudo-data (for more information on the non-gaussian model see Seeger & Bouchard, 2012).
}
\details{
fill this
}
