% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSamples.R
\name{clusterSamples}
\alias{clusterSamples}
\title{Hierarchical clustering on samples based on latent factors}
\usage{
clusterSamples(object, factors = "all", anno_df = NULL, main = NULL, ...)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{factors}{factors to use for the clustering}

\item{anno_df}{annotation data frame that will be passed to pheatmap}

\item{main}{title of the plot}
}
\value{
A hclust object containing the clustering of the samples
}
\description{
fill this
}
\details{
If only one factor (...). Plots heatmap
}
