% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSamples.R
\name{clusterSamples}
\alias{clusterSamples}
\title{Cluster samples in a MOFA model based on latent factors}
\usage{
clusterSamples(object, factors = "all", anno_df = NULL, main = NULL)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{factors}{factors to use for the clustering}
}
\value{
A hclust object containing the clustering of the samples
}
\description{
fill this
}
\details{
asd
}
