% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSamples.R
\name{clusterSamples}
\alias{clusterSamples}
\title{K-means clustering on samples based on latent factors}
\usage{
clusterSamples(object, k, factors = "all", ...)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{k}{number of clusters}

\item{factors}{character vector with the factor name(s), or numeric vector with the index of the factor(s) to use. Default is 'all'\#' @param factors character vector with the factor name(s), or numeric vector with the index of the factor(s) to use. Default is 'all'}

\item{...}{other arguments that can be passed to the kmeans function}
}
\value{
output from \code{\link{kmeans}} function
}
\description{
latent factors are continuous in nature but they can be used to predict clusters of samples, similar to what the iCluster model does (Shen, 2009). \cr
The clustering can be performed in a single factor, which is equivalent to setting a manual threshold; 
or using multiple factors, where you combine multiple sources of variation. \cr
}
\details{
In some cases, due to model technicalities, samples can have missing values in the latent factor space. In such a case, these samples are currently removed.
}
