% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.R
\name{compareModels}
\alias{compareModels}
\title{Compare the latent factors of different \code{\link{MOFAmodel}}}
\usage{
compareModels(ModelList, comparison = "all")
}
\arguments{
\item{ModelList}{a list containing \code{\link{MOFAmodel}} objects.}

\item{comparison}{tye of comparison either 'pairwise' or 'all'}
}
\value{
Plots a heatmap of correlation of Latent Factors in all models when 'comparison' is 'all'. 
Otherwise, for each pair of models, a seperate heatmap is produced comparing one model againt the other.
The corresponding correlation matrix or list or pairwise correlation matrices is returned
}
\description{
Different objects of \code{\link{MOFAmodel}} are compared in terms of correlation between 
their latent factors. The correlation is calculated only on those samples which are present in all models 
used for a comparison.
}
\details{
asd
}
\references{
fill this
}
