% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPlot.R
\name{CorrplotLFvsPC}
\alias{CorrplotLFvsPC}
\title{Correlate latent factors to principal components on single views}
\usage{
CorrplotLFvsPC(model, views = "all", noPCs = 5, method = "svd")
}
\arguments{
\item{views}{fill}

\item{noPCs}{fill}

\item{method}{fill, can be svd, ppca, bpca}

\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
Correlation matrix of latent factors versus principal components
}
\description{
fill this
}
\details{
fill
}
