% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMOFAobject.R
\name{createMOFAobject}
\alias{createMOFAobject}
\title{Initialize a \code{\link{MOFAmodel}} object with multi-omic data}
\usage{
createMOFAobject(data)
}
\arguments{
\item{data}{either MultiAssayExperiment or list of matrices with features as rows and samples as columns}
}
\value{
an untrained MOFA model object
}
\description{
Method to initialize a MOFAmodel object
}
