% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{Expectations}
\alias{Expectations}
\alias{Expectations}
\alias{Expectations<-}
\alias{Expectations,MOFAmodel-method}
\title{Expectations: set and retrieve expectations}
\usage{
Expectations(object)

.Expectations(object) <- value

\S4method{Expectations}{MOFAmodel}(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
list of expectations
}
\description{
Function to set and retrieve model expectations.
}
\examples{
# load existing object
filepath <- system.file("extdata", "scMT_model.hdf5", package = "MOFAtools")
MOFAobject <- loadModel(filepath)
# get training statistics
stats <- TrainStats(MOFAobject)
# load a trained MOFAmodel object
filepath <- system.file("extdata", "CLL_model.hdf5", package = "MOFAtools")
MOFAobject <- loadModel(filepath)
# get expectations 
expectations <- Expectations(MOFAobject)
}
