% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{Expectations}
\alias{Expectations}
\alias{Expectations,MOFAmodel-method}
\title{Expectations: set and retrieve expectations}
\usage{
Expectations(object)

\S4method{Expectations}{MOFAmodel}(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
list of expectations
}
