% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFactors.R
\name{FactorHistPlot}
\alias{FactorHistPlot}
\title{Visualize histogram of one latent factor}
\usage{
FactorHistPlot(object, factor, group_by = NULL, group_names = "",
  alpha = 0.6, binwidth = NULL, xlabel = "", showMissing = FALSE)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{factor}{character vector with the factor name or numeric vector with the index of the factor.}

\item{group_by}{specifies groups used to color the samples of the histogram. This can be either a character giving the name of a feature, or the name of a covariate if using a MultiAssayExperiment class, or a vector of same length as number of samples.}

\item{group_names}{name for groups (usually only used if groups is not a character itself)}

\item{alpha}{transparency parameter}

\item{binwidth}{binwidth for histogram (default is NULL, which uses ggplot's default calculation)}

\item{xlabel}{define x-axis label (default is NULL)}

\item{showMissing}{boolean indicating whether to remove sample for which 'group_by' is missing (default is FALSE)}
}
\value{
ggpplot object
}
\description{
generate a histogram of the sample values in a given latent factor
}
\details{
One of the first steps for the annotation of factors is to visualise and group/color them using known covariates such as phenotypic or clinical data.
This method generates a histogram of the sample values in a given latent factor. Note that, similar to Principal Component Analysis, the factor values should be interpreted in a relative manner.
Similar functions are \code{\link{FactorsScatterPlot}} for doing scatter plots and \code{\link{FactorBeeswarmPlot}} for doing Beeswarm plots
}
