% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPlot.R
\name{FactorsCorPlot}
\alias{FactorsCorPlot}
\title{Visualize correlation matrix between the latent variables}
\usage{
FactorsCorPlot(object, method = "pearson", ...)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{method}{a character string indicating which correlation coefficient is to be computed: pearson (default), kendall, or spearman.}

\item{...}{arguments passed to \code{corrplot}}
}
\value{
symmetric matrix with the correlation coefficient between factors
}
\description{
plot the correlation matrix between the latent variables. The model encourages factors to be uncorrelated, but it is nos hard constraint such as in PCA.
Ideally all learnt factors should be uncorrelated to make interpretation easier, but correlations can happen, particularly with large number factors or with non-linear relationship.
If you have too many correlated factors try to train the model again reducing the number of factors.
}
