% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPlot.R
\name{FactorsCorPlot}
\alias{FactorsCorPlot}
\title{Visualize correlation matrix between the latent variables}
\usage{
FactorsCorPlot(object, method = "pearson", ...)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{method}{a character string indicating which correlation coefficient is to be computed: pearson (default), kendall, or spearman.}

\item{...}{arguments passed to \code{corrplot}}
}
\value{
fill this
}
\description{
fill this
}
\details{
asd
}
