% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPlot.R
\name{FactorsScatterPlot}
\alias{FactorsScatterPlot}
\title{Faced scatterplot between latent variables}
\usage{
FactorsScatterPlot(object, z_order = NULL, title = "")
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{z_order}{...}

\item{title}{fill this}
}
\value{
fill this
}
\description{
fill this
}
\details{
asd
}
