% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{featureNames}
\alias{featureNames}
\alias{featureNames<-}
\alias{featureNames,MOFAmodel-method}
\alias{featureNames<-,MOFAmodel,list-method}
\title{featureNames: set and retrieve feature names}
\usage{
featureNames(object)

featureNames(object) <- value

\S4method{featureNames}{MOFAmodel}(object)

\S4method{featureNames}{MOFAmodel,list}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{value}{list of character vectors with the feature names for each view}
}
\value{
list of character vectors with the feature names for each view
}
\author{
Ricard Argelaguet
}
