% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{FeatureSetEnrichmentAnalysis}
\alias{FeatureSetEnrichmentAnalysis}
\title{Feature Set Enrichment Analysis}
\usage{
FeatureSetEnrichmentAnalysis(model, view, feature.sets, factors = "all",
  local.statistic = c("loading", "cor", "z"),
  transformation = c("abs.value", "none"), global.statistic = c("mean.diff",
  "rank.sum"), statistical.test = c("parametric", "cor.adj.parametric",
  "permutation"), nperm = 100, min.size = 10, cores = 1,
  p.adj.method = "BH", alpha = 0.1)
}
\arguments{
\item{model}{a \code{\link{MOFAmodel}} object.}

\item{view}{name of the view}

\item{feature.sets}{Data structure that holds feature set membership information. Must be either a binary membership matrix (rows are feature sets and columns are features) or a list of feature set member indexes (see vignette).}

\item{factors}{either name (character vector) or indices (numeric vector) of the latent variables for which enrichment should be computed (default="all")}

\item{local.statistic}{The feature statistic used to quantify the association between each genomic variable and each PC. Must be one of the following: loading (default), cor, z.}

\item{transformation}{Optional transformation to apply to the feature-level statistics. Must be one of the following "none" or "abs.value" (default).}

\item{global.statistic}{The feature set statisic computed from the feature statistics. Must be one of the following: "mean.diff" (default) or "rank.sum".}

\item{statistical.test}{The statistical test used to compute the significance of the feature set statistics under a competitive null hypothesis.
Must be one of the following: "parametric" (default), "cor.adj.parametric", "permutation".}

\item{nperm}{Number of permutations to perform. Only relevant if statistical.test is set to "permutation".}

\item{min.size}{Minimum size of a feature set}

\item{cores}{Number of cores to run the permutation analysis in parallel. Only relevant if statistical.test is set to "permutation".}

\item{p.adj.method}{Method to adjust p-values factor-wise for multiple testing. Can be any method in p.adjust.methods(). Default uses Benjamini-Hochberg procedure.}
}
\value{
a list with two matrices, one for the feature set statistics and the other for the pvalues. Rows are feature sets and columns are latent variables.
}
\description{
Method to perform enrichment analysis for feature sets from a given view in the latent factors  \cr
Here we use a slightly modified version of the \link{PCGSE} package.
}
\details{
fill this
}
