% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getCovariates}
\alias{getCovariates}
\title{getCovariates}
\usage{
getCovariates(object, covariate)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{covariate}{names of the covariate}
}
\description{
This function extracts covariates from the \code{colData}
 in the input \code{MultiAssayExperiment} object. \cr
Note that if you did not use \code{MultiAssayExperiment} to create
 your \code{\link{createMOFAobject}}, this function will not work.
}
