% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getCovariates}
\alias{getCovariates}
\title{wraper to extract covariates from the MultiAssayExperiment stored in the InputData slot}
\usage{
getCovariates(object, names)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{names:}{names of the covariates}
}
\description{
to-fill
}
