% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getCovariates}
\alias{getCovariates}
\title{Extract covariates from the model}
\usage{
getCovariates(object, names)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{names:}{names of the covariates}
}
\description{
this function extracts the covariates defined by the user. 
Note that (for now) the model can be run with known covariates only if they are stored in the colData of the MultiAssayExperiment input object. See \code{\link{createMOFAmodel}}.
}
