% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMOFA.R
\name{getDefaultModelOpts}
\alias{getDefaultModelOpts}
\title{Get default model options}
\usage{
getDefaultModelOpts(object)
}
\arguments{
\item{object}{untrained MOFA object to get model options for}
}
\value{
list with default model options
}
\description{
Function to obtain default model options
}
\details{
The model options are the following: \cr
likelihood: character vector with data likelihoods per view, 'gaussian' for continuous data, 'bernoulli' for binary data and 'poisson' for count data. \cr
numFactors: numeric indicating the initial number of factors. If you have no prior expectation, we recommend this to be larger than 10. Default is 25. \cr
learnIntercept: logical indicating whether to learn the intercept (the means) per feature. This is always recommended, particularly if you have non-gaussian likelihoods. Default is TRUE. \cr
sparsity: logical indicating whether to use sparsity. This is always recommended, as it will make the loadings more interpretable. Default is TRUE. \cr
covariates: not implemented yet.
}
