% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMOFA.R
\name{getDefaultModelOpts}
\alias{getDefaultModelOpts}
\title{getDefaultModelOpts: Get default model options}
\usage{
getDefaultModelOpts(object, silent = T)
}
\arguments{
\item{object}{untrained MOFA object to get model options for}

\item{silent}{boolean whether to print warnings}
}
\value{
list with training options
}
\description{
Function to obtain default model options
}
\details{
fill this
}
