% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMOFA.R
\name{getDefaultTrainOpts}
\alias{getDefaultTrainOpts}
\title{Get default training options}
\usage{
getDefaultTrainOpts()
}
\value{
list with default training options
}
\description{
Function to obtain default training options
}
\details{
The training options are the following: \cr
maxiter: numeric indicating the maximum number of iterations. \cr
tolerance: numeric indicating the convergence threshold based on the change in Evidence Lower Bound. 
For quick exploration we recommend this to be around 1.0, and for a thorough training we recommend a value of 0.01. \cr
learnFactors: logical indicating whether to learn the number of factors based on the the minimum fraction of variance explained? \cr
DropFactorThreshold: numeric indicating the threshold on fraction of variance explained to drop a factor. 
That is, factors explaining less than 'DropFactorThreshold' fraction of variance (in all views) will be dropped. \cr
verbose: logical indicating whether to generate a verbose output? \cr
}
