% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMOFA.R
\name{getDefaultTrainOpts}
\alias{getDefaultTrainOpts}
\title{getDefaultTrainOpts: Get default training options}
\usage{
getDefaultTrainOpts(silent = T)
}
\value{
list with training options
}
\description{
Function to obtain default training options
}
\details{
fill this
}
