% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getExpectations}
\alias{getExpectations}
\title{wraper to fetch particular expectations from the model}
\usage{
getExpectations(object, variable, expectation, as.data.frame = F)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{variable:}{variable name ('SW')}

\item{expectation:}{expectation name ('E')}

\item{as.data.frame:}{output the result as a long data frame?}
}
\description{
to-fill
}
