% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getFactors}
\alias{getFactors}
\title{Extract the latent factors from the model}
\usage{
getFactors(object, factors = "all", as.data.frame = FALSE,
  include_intercept = TRUE)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{factors}{character vector with the factor name(s), or numeric vector with the factor index(es) (here 0 corresponds to the intercept factor if included in the model). Default is "all".}

\item{as.data.frame}{boolean indicating whether to return a long data frame instead of a matrix, default is FALSE.}

\item{include_intercept}{boolean indicating where to include the intercept term of the model, if present. Default is TRUE.}
}
\value{
By default returns the latent factor matrix of dimensionality (N,K), where N is number of samples and K is number of factors. \cr
Alternatively, if as.data.frame is TRUE, returns a long-formatted data frame with columns (sample,factor,value).
}
