% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getParameters}
\alias{getParameters}
\title{wraper to fetch particular parameters from the model}
\usage{
getParameters(object, variable, parameter)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{variable:}{variable name ('SW')}

\item{parameter_name:}{parameter name ('E')}
}
\description{
to-fill
}
