% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getWeights}
\alias{getWeights}
\title{Extract the weights from the model}
\usage{
getWeights(object, views = "all", factors = "all", as.data.frame = F)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{views}{character vector with the view name(s), or numeric vector with the view index(es), default is "all".}

\item{factors}{character vector with the factor name(s) or numeric vector with the factor index(es), default is "all".}

\item{as.data.frame}{boolean indicating whether to return a long data frame instead of a list of matrices, default is FALSE.}
}
\value{
by default returns a list where each element is a loading matrix with dimensionality (D,K) where D is the number of features in this view and K is the number of factors.
Alternatively, if as.data.frame is TRUE, returns a long-formatted data frame with columns (view,feature,factor,value)
}
\details{
TO-DO: INCLUDE FEATURES
}
