% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{Heatmap_FeatureSetEnrichmentAnalysis}
\alias{Heatmap_FeatureSetEnrichmentAnalysis}
\title{Heatmap of Feature Set Enrichment Analysis results}
\usage{
Heatmap_FeatureSetEnrichmentAnalysis(p.values, threshold = 0.05, ...)
}
\arguments{
\item{p.values}{output of \link{FeatureSetEnrichmentAnalysis} function. A data frame of p.values where rows are gene sets and columns are latent variables}

\item{threshold}{p.value threshold to filter out feature sets. If a feature set has a p.value lower than 'threshold'}

\item{...}{Parameters to be passed to pheatmap function}
}
\value{
nothing
}
\description{
Heatmap of the Feature Set Enrichment Analyisis results
}
\details{
fill this
}
