% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{histPlot}
\alias{histPlot}
\title{Visualize histogram of one latent variable}
\usage{
histPlot(object, id, xlabel = NULL, groups = NULL, alpha = 1,
  binwidth = NULL)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{id}{latent variable}
}
\value{
fill this
}
\description{
fill this
}
\details{
asd
}
