% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{histPlot}
\alias{histPlot}
\title{Visualize histogram of one latent variable}
\usage{
histPlot(object, factor, xlabel = NULL, fill = NULL, name_fill = "",
  alpha = 0.6, binwidth = NULL, showNA = F)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{id}{name of latent variable to plot}
}
\value{
fill this
}
\description{
fill this
}
\details{
asd
}
\references{
fill this
}
