% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{ImputedData}
\alias{ImputedData}
\alias{ImputedData,MOFAmodel-method}
\title{ImputedData: set and retrieve training data}
\usage{
ImputedData(object)

\S4method{ImputedData}{MOFAmodel}(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
list of numeric matrices that contain the training data
}
