% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissing.R
\name{imputeMissing}
\alias{imputeMissing}
\title{Impute missing value from a fitted MOFA model}
\usage{
imputeMissing(model, viewnms = "all")
}
\arguments{
\item{viewnms}{character vector containing the names of views to be imputed (default: all)}

\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
List of imputed data, each list element corresponding to specified views.
}
\description{
fill this
}
\details{
asd
}
