% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissing.R
\name{imputeMissing}
\alias{imputeMissing}
\title{Impute missing values from a fitted MOFA model}
\usage{
imputeMissing(object, views = "all", factors = "all", type = c("inRange",
  "response", "link"))
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{views}{vector containing the names of views (character) or index of views (numeric) to be imputed (default: "all")}

\item{factors}{vector with the factors indices (numeric) or factor names (character) to use (default is "all")}

\item{type}{type of imputations returned. By default values are imputed using "inRange". "response" gives mean for gaussian and poisson and probabilities for bernoulli , 
"link" gives the linear predictions, "inRange" rounds the fitted values from "terms" for integer-valued distributions to the next integer.}
}
\value{
List of imputed data, each list element corresponding to specified views.
}
\description{
This function uses the latent factors and the weights infered from MOFA to impute missing values in the input views.
}
\details{
asd
}
\references{
fill this
}
