% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{InputData}
\alias{InputData}
\alias{InputData}
\alias{inputData<-}
\alias{InputData}
\title{Set and retrieve input data}
\usage{
InputData(object)

.InputData(object) <- value

\S4method{InputData}{MOFAmodel}(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
A MultiAssayExperiment object or a list of matrices, one per view,
 containing the input data before beign parsed for training
}
\description{
Function to set and retrieve input data.
}
\examples{
data("scMT_data")
MOFAobject <- createMOFAobject(scMT_data)
InputData(MOFAobject)
}
