% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSEA.R
\name{LinePlot_FeatureSetEnrichmentAnalysis}
\alias{LinePlot_FeatureSetEnrichmentAnalysis}
\title{Line plot of Feature Set Enrichment Analysis results}
\usage{
LinePlot_FeatureSetEnrichmentAnalysis(object, fsea.out, factor, alpha = 0.1,
  max.pathways = 25, adjust = T)
}
\arguments{
\item{fsea.out}{output of \link{FeatureSetEnrichmentAnalysis} function}

\item{factor}{Factor for which to show wnriched pathways in the lineplot}

\item{alpha}{p.value threshold to filter out feature sets}

\item{max.pathways}{maximum number of enriched pathways to display}

\item{adjust}{use multiple testing correction}
}
\value{
nothing
}
\description{
Line plot of the Feature Set Enrichment Analyisis results for a specific latent variable
}
\examples{
# Example on the CLL data
filepath <- system.file("extdata", "CLL_model.hdf5", package = "MOFAtools")
MOFA_CLL <- loadModel(filepath)
# perform FeatureSetEnrichmentAnalysis on mRNA data for all factors using the Reactome gene sets
data("reactomeGS")
fsea.out <- FeatureSetEnrichmentAnalysis(MOFA_CLL, view="mRNA", feature.sets=reactomeGS)
# top 10 enriched pathwyas on factor 5:
LinePlot_FeatureSetEnrichmentAnalysis(MOFA_CLL, fsea.out, factor=5,  max.pathways=10)
}
