% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModel.R
\name{loadModel}
\alias{loadModel}
\title{loadModel: loading a trained MOFA model from python output}
\usage{
loadModel(file, model = NULL)
}
\arguments{
\item{file}{an hd5f model generated by the MOFA python package.}
}
\value{
a \code{\link{MOFAmodel}} model.
}
\description{
Method to load a trained MOFA model from a hdf5 file. \cr
Models have to be trained and saved using the corresponding Python package.
}
\details{
fill this
}
