% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{ModelOptions}
\alias{ModelOptions}
\alias{ModelOptions}
\alias{ModelOptions<-}
\alias{ModelOptions,MOFAmodel-method}
\title{ModelOptions: set and retrieve Model options}
\usage{
ModelOptions(object)

.ModelOptions(object) <- value

\S4method{ModelOptions}{MOFAmodel}(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
list of model options
}
\description{
Function to set and retrieve model options.
}
\examples{
# load existing object
filepath <- system.file("extdata", "scMT_model.hdf5", package = "MOFAtools")
MOFAobject <- loadModel(filepath)
# get ModelOptions
ModelOptions(MOFAobject)
}
