% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MOFAmodel}
\alias{MOFAmodel}
\alias{MOFAmodel-class}
\title{Class to store a Multi-Omics Factor Analysis (MOFA) model}
\description{
The \code{MOFAmodel} is an S4 class used to store all relevant data to analyse a MOFA model.
}
