% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotDataHeatmap}
\alias{plotDataHeatmap}
\title{plotDataHeatmap: plot heatmap of selected features}
\usage{
plotDataHeatmap(object, view, factor, features = 50, includeWeights = FALSE,
  transpose = FALSE, imputed = FALSE, ...)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{view}{character vector with the view name, or numeric vector with the index of the view.}

\item{factor}{character vector with the factor name, or numeric vector with the index of the factor.}

\item{features}{if an integer, the total number of features to plot, based on the absolute value of the loading (50 by default).
If a character vector, a set of manually-defined features to plot.}

\item{includeWeights}{boolean indicating whether to include the weight of each feature as an extra annotation in the heatmap (FALSE by default).}

\item{transpose}{boolean indicating whether to transpose the output heatmap (FALSE by default, which corresponds to features as rows and samples as columns)}

\item{imputed}{boolean indicating whether to use the imputed data (FALSE by default)}

\item{...}{further arguments that can be passed to \code{\link[pheatmap]{pheatmap}}}
}
\description{
Function to plot a heatmap for selected features, usually the ones with highest loadings in a given latent factor.
}
\details{
One of the first steps for the annotation of factors is to visualise the corresponding loadings using \code{\link{plotWeights}} or \code{\link{plotTopWeights}}. \cr
These methods show you which features drive the heterogeneity of each factor. However, one might also be interested in directly visualising the heterogeneity in the original data, rather than looking at "abstract" weights. \cr
This method plots a heatmap for selected features (by default the top ones with highest loading), which should reveal the underlying pattern of the data that is captured by the latent factor. \cr
A similar function for doing scatterplots rather than heatmaps is \code{\link{plotDataScatter}}.
}
