% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotDataScatter}
\alias{plotDataScatter}
\title{plotDataScatter: scatterplot of features against latent factors}
\usage{
plotDataScatter(object, view, factor, features = 10, color_by = NULL,
  shape_by = NULL)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{view}{character vector with a view name, or numeric vector with the index of the view.}

\item{factor}{character vector with a factor name, or numeric vector with the index of the factor.}

\item{features}{if an integer, the total number of features to plot (10 by default). If a character vector, a set of manually-defined features.}

\item{color_by}{specifies groups or values used to color points. This can be either a character giving the name of a feature, or the name of covariate from the MultiAssayExperiment object, or a factor vector of same length as the number of samples.}

\item{shape_by}{specifies groups or values used to shape points, same behaviour as 'color_by'}
}
\description{
Function to do a scatterplot of feature(s) against a latent factor.
}
\details{
One of the first steps for the annotation of factors is to visualise the loadings using \code{\link{plotWeights}} or \code{\link{plotTopWeights}}. \cr
These methods show you which features drive the heterogeneity of each factor. However, one might also be interested in visualising the heterogeneity in the original data, rather than looking at "abstract" weights. \cr
This method generates scatterplots of features against factors, so that you can observe the association between them. \cr
A similar function for doing heatmaps rather than scatterplots is \code{\link{plotDataHeatmap}}.
}
