% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotTilesData}
\alias{plotTilesData}
\title{plotTilesData: tile plot of data availability for each sample}
\usage{
plotTilesData(object, colors = NULL)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{colors}{a vector specifying the colors per view.}
}
\description{
Function to do a tile plot showing the structure of the multi-omics input data
}
\details{
This function is helpful to get an overview of the structure of the training data used for MOFA. 
It shows the number of samples, the number of views, the number of features, and the structure of missing values.
In particular, it is useful to visualise incomplete data sets, where some samples are missing subsets of assays.
}
