% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWeights.R
\name{plotTopWeights}
\alias{plotTopWeights}
\title{plotTopWeights: visualise the top weights for a certain factor and view}
\usage{
plotTopWeights(object, view, factor, nfeatures = 5, abs = TRUE,
  scale = TRUE, sign = "both")
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{view}{character vector with the view name or numeric vector with the index of the view.}

\item{factor}{character vector with the factor name or numeric vector with the index of the factor.}

\item{nfeatures}{number of top features to display.}

\item{abs}{take absolute value of the weights? (default=TRUE)}

\item{scale}{scales all loading by absolute value of maximal loading}

\item{sign}{can be 'positive', 'negative' or 'both' to show only positive, negative or all weigths, respectively}
}
\description{
The first step to annotate factors is to visualise the corresponding feature loadings. \cr
This function zooms and displays the features with highest loading in a given latent factor and view. \cr
In contrast, the function \code{showAllWeights} displays the entire distribution of weights.
}
