% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWeights.R
\name{plotWeights}
\alias{plotWeights}
\title{Plot Weights}
\usage{
plotWeights(object, view, factor, nfeatures = 10, abs = FALSE,
  manual = NULL, color_manual = NULL, scale = TRUE)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{view}{character vector with the view name, or numeric vector with the index of the view to use.}

\item{factor}{character vector with the factor name, or numeric vector with the index of the factor to use.}

\item{nfeatures}{number of top features to label.}

\item{abs}{logical indicating whether to use the absolute value of the weights.}

\item{manual}{A nested list of character vectors with features to be manually labelled.}

\item{color_manual}{a character vector with colors, one for each element of 'manual'}

\item{scale}{logical indicating whether to scale all loadings from 0 to 1.}
}
\description{
An important step to annotate factors is to visualise the corresponding feature loadings. \cr
This function plots all loadings for a given latent factor and view, labeling the top ones. \cr
In contrast, the function \code{\link{plotTopWeights}} displays only the top features with highest loading.
}
\details{
The weights of the features within a view are relative and they should not be interpreted in an absolute scale.
For interpretability purposes we always recommend to scale the weights with \code{scale=TRUE}.
}
\examples{
# Example on the CLL data
filepath <- system.file("extdata", "CLL_model.hdf5", package = "MOFAtools")
MOFA_CLL <- loadModel(filepath)
plotWeights(MOFA_CLL, view="Mutations", factor=1)
plotWeights(MOFA_CLL, view="Mutations", factor=1,
  manual=list("IGHV", c("TP53", "del17p13")), color_manual=c("blue", "red"))

# Example on the scMT data
filepath <- system.file("extdata", "scMT_model.hdf5", package = "MOFAtools")
MOFA_scMT <- loadModel(filepath)
plotWeights(MOFA_scMT, view="RNA expression", factor=1)
plotWeights(MOFA_scMT, view="RNA expression", factor=1, nfeatures=15)
}
