% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Do predictions using a fitted MOFA model}
\usage{
predict(object, views = "all", factors = "all", type = c("inRange",
  "response", "link"), include_intercept = TRUE)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{views}{character vector with the view name(s), or numeric vector with the view index(es), default is "all".}

\item{factors}{character vector with the factor name(s) or numeric vector with the factor index(es), default is "all".}

\item{type}{type of prediction returned. "response" gives mean for gaussian and poisson, and probabilities for bernoulli , 
"link" gives the linear predictions, "inRange" rounds the fitted values from "terms" for integer-valued distributions to the next integer. Default is "inRange".}
}
\value{
List with data predictions, each element corresponding to a view.
}
\description{
This function uses the latent factors and the weights to do predictions in the input data
}
\details{
the denoised and condensed low-dimensional representation of the data captures the main sources of heterogeneity of the data. 
These representation can be used to do predictions using the equation Y = WX. This is the key step underlying imputation, see \code{\link{imputeMissing}} and Methods section of the article.
}
