% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMOFA.R
\name{prepareMOFA}
\alias{prepareMOFA}
\title{prepareMOFA: Prepare an untrained MOFA object for running MOFA}
\usage{
prepareMOFA(object, dir, ModelOptions = NULL, TrainOptions = NULL,
  outFile = "MOFAout", k = 10, MOFAdir)
}
\arguments{
\item{object}{an untrained MOFA object}

\item{dir}{directory to store .txt and .sh files in}

\item{ModelOptions}{list of ModelOptions (see getDefaultModelOpts for what options can be set here). If none specified, default options are used.}

\item{TrainOptions}{list of TrainOptions (see getDefaultTrainOptions for what options can be set here). If none specified, default options are used.}

\item{outFile}{name of output file from Python MOFA}

\item{k}{number of latent factors to start with (default = 10)}

\item{MOFAdir}{directory of the MOFA Pyhton package installation}
}
\value{
a untrained MOFA object with specified ModelOpts and TrainOpts
}
\description{
Function to set the training and model options, produces .txt files that are used for python as input and 
creates an .sh file for calling MOFA with the specified options from the command line. These files are all stored in the specified directory.
}
\details{
fill this
}
