% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMOFA.R
\name{prepareMOFA}
\alias{prepareMOFA}
\title{prepare a MOFA model for training}
\usage{
prepareMOFA(object, DirOptions, DataOptions = NULL, ModelOptions = NULL,
  TrainOptions = NULL)
}
\arguments{
\item{object}{an untrained \code{\link{MOFAmodel}}}

\item{DirOptions}{list with Input/Output options, which must contain a 'dataDir' element where temporary text files will be stored 
and an 'outFile' with hdf5 extension where the final model will be stored.}

\item{DataOptions}{list of DataOptions (see \code{\link{getDefaultDataOptions}} details). 
If NULL, default data options are used.}

\item{ModelOptions}{list of ModelOptions (see \code{\link{getDefaultModelOptions}} for details). 
If NULL, default model options are used.}

\item{TrainOptions}{list of TrainOptions (see \code{\link{getDefaultTrainOptions}} for details). 
If NULL, default training options are used.}
}
\value{
Returns an untrained \code{\link{MOFAmodel}} with specified data, model and training options
}
\description{
Function to prepare a \code{\link{MOFAmodel}} object for training. It requires defining data, model and training options.
}
\examples{
# Generate a random data set
gaussian_data <- matrix(rnorm(100,mean=0,sd=1),10,10)
poisson_data <- matrix(rpois(100, lambda=1),10,10)
data <- list("gaussian_view"=gaussian_data, "poisson_view"=poisson_data)
# Create a MOFA object
MOFAobject <- createMOFA(data)
# Define I/O options
DirOptions <- list("dataDir" = tempdir(), "outFile" = tempfile())
# Define Data Options
DataOptions <- getDefaultDataOptions()
# Define Model Options
ModelOptions <- getDefaultModelOptions(MOFAobject)
ModelOptions$likelihood <- c("gaussian","poisson")
ModelOptions <- getDefaultModelOptions(MOFAobject)
# Define Training Options
TrainOptions <- getDefaultTrainOptions()
TrainOptions$maxiter <- 100
# Prepare MOFA object for training
MOFAobject <- prepareMOFA(MOFAobject, DirOptions, DataOptions, ModelOptions, TrainOptions)
# train a MOFA model
MOFAobject <- runMOFA(MOFAobject, DirOptions)
}
