% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{qualityControl}
\alias{qualityControl}
\title{qualityControl}
\usage{
qualityControl(object, verbose = FALSE)
}
\arguments{
\item{object}{a trained \code{\link{MOFAmodel}} object.}

\item{verbose}{logical indicating whether to generate a verbose output.}
}
\value{
none
}
\description{
Function to do quality control on a \code{\link{MOFAmodel}} object.
}
\examples{
# load a trained MOFAmodel object
filepath <- system.file("extdata", "CLL_model.hdf5", package = "MOFAtools")
MOFAobject <- loadModel(filepath)
# do quality control
qualityControl(MOFAobject, verbose=TRUE)
}
