% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{qualityControl}
\alias{qualityControl}
\title{qualityControl}
\usage{
qualityControl(object, verbose = F)
}
\arguments{
\item{object}{a trained \code{\link{MOFAmodel}} object.}

\item{verbose}{logical indicating whether to generate a verbose output.}
}
\description{
Function to do quality control on a \code{\link{MOFAmodel}} object.
}
