% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMOFA.R
\name{runMOFA}
\alias{runMOFA}
\title{train a MOFA model}
\usage{
runMOFA(object)
}
\arguments{
\item{object}{an untrained \code{\link{MOFAmodel}} object}
}
\value{
a trained \code{\link{MOFAmodel}} object
}
\description{
Function to train an untrained \code{\link{MOFAmodel}} object.
}
\details{
In this step the R package is calling the \code{mofa} Python package, where the the training is performed. \cr
The interface with Python is done with the \code{\link{reticulate}} package. 
If you have several versions of Python installed and Rstudio is not detecting the correct one, you can change it using
\code{reticulate::use_python}. \cr
This module is in beta testing so please, read our FAQ for troubleshooting and report any problems.
}
\examples{
data("CLL_data")
#create and prepare the MOFaobject before using runMOFA
MOFAobject <- createMOFAobject(CLL_data)
MOFAobject <- prepareMOFA(MOFAobject)
# runMOFA takes some time for model fitting
\dontrun{
# MOFAobject <- runMOFA(MOFAobject)}
}
