% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMOFA.R
\name{runMOFA}
\alias{runMOFA}
\title{runMOFA:}
\usage{
runMOFA(object, DirOptions)
}
\arguments{
\item{object}{an untrained \code{\link{MOFAmodel}}}

\item{DirOptions}{list with I/O options, should contain at least 'dataDir' where the input matrices as stored as .txt files and 'outFile' where the model is going to be stored as a .hdf5 file}
}
\value{
a trained \code{\link{MOFAmodel}}
}
\description{
train a \code{\link{MOFAmodel}}
}
