% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPairs}
\alias{scatterPairs}
\title{Visualize scatterplot of all latent variables in a pair-wise grid}
\usage{
scatterPairs(object, factors = "all", showMissing = T, color_by = NULL,
  color_name = NULL, colorLegend = F, shape_by = NULL,
  shape_name = NULL, shapeLegend = F)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{showMissing}{boolean wether to include missing values in color_by}

\item{color_by}{vector of length N with discrete sample groups to color by}

\item{color_name}{name of the factor used to color_by (only relevant if color_by is not NULL)}
}
\value{
fill this
}
\description{
fill this
}
\details{
asd
}
\references{
fill this
}
