% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPairs}
\alias{scatterPairs}
\title{Visualize scatterplot of all latent variables in a pair-wise grid}
\usage{
scatterPairs(object, factors = "all", colour_by = NULL, showMissing = T,
  colour_name = NULL)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{colour_by}{discrete vector of length N (samples in object) to colour samples by}

\item{showMissing}{boolean wether to include NA values of coulour_by}
}
\value{
fill this
}
\description{
fill this
}
\details{
asd
}
