% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showData.R
\name{showDataHeatmap}
\alias{showDataHeatmap}
\title{showDataHeatmap: heatmap of the observed data of the most relevant features for a given factor}
\usage{
showDataHeatmap(model, view, factor, nfeatures = 50, main = NULL, ...)
}
\arguments{
\item{model}{a fitted MOFA model}

\item{view}{name of view}

\item{factor}{factor}

\item{nfeatures}{total number of features, sorted by weight (default=???)}

\item{...}{further arguments that can be passed to pheatmap}
}
\value{
fill this
}
\description{
Function to plot a heatmap of the original data using the most relevant features of a given latent variables
}
\details{
fill this
}
