% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showData.R
\name{showDataHeatmap}
\alias{showDataHeatmap}
\title{showDataHeatmap: plot heatmap of the observed data for the top weighted features}
\usage{
showDataHeatmap(object, view, factor, features = 50, plotWeights = FALSE,
  transpose = FALSE, ...)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{view}{character vector with the view name or numeric vector with the index of the view.}

\item{factor}{character vector with the factor name or numeric vector with the index of the factor.}

\item{features}{if an integer, the total number of features to plot sorted by the corresponding weight.
If a character vector, the manually-defined features to plot by the given order.}

\item{...}{further arguments that can be passed to pheatmap}

\item{plotWeights:}{boolean indicating whether to include the weight of each feature in the heatmap.}

\item{transpose:}{boolean indicating whether to transpose the output heatmap (by default features as rows and samples as columns)}
}
\description{
Function to plot a heatmap of the original data using the most relevant features for a given latent variable.
}
\details{
fill this
}
