% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showData.R
\name{showDataScatter}
\alias{showDataScatter}
\title{showDataScatter: scatterplot of the observed data for the most relevant features for a given factor}
\usage{
showDataScatter(model, view, factor, nfeatures = 50, colour_by = NULL,
  shape_by = NULL, xlabel = "", ylabel = "")
}
\arguments{
\item{model}{a fitted MOFA model}

\item{view}{name of view}

\item{factor}{factor}

\item{nfeatures}{total number of features (default=???)}

\item{shape_by}{vector}

\item{color_by}{vector}
}
\value{
ggplot2 model
}
\description{
Function to plot a scatterplot of the observed values for the most relevant features for a given factor
}
\details{
fill this
}
