% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showWeights.R
\name{showWeightHeatmap}
\alias{showWeightHeatmap}
\title{showWeightHeatmap: show heatmap of the factor loadings in a specific view}
\usage{
showWeightHeatmap(model, view, features = "all", factors = "all",
  RemoveIntercept = T, main = NULL, color = NULL, breaks = NULL, ...)
}
\arguments{
\item{model}{a MOFA model}

\item{view}{view name}

\item{features}{name of the features to include (default: "all")}

\item{factors}{name of the factors to include (default: "all")}

\item{RemoveIntercept}{boolean wether to include the intercept factor if present in the model (by default not included)}

\item{main}{asd}

\item{color}{asd}

\item{breaks}{asd}

\item{...}{further arguments that can be passed to pheatmap}
}
\value{
fill this
}
\description{
Function to visualize the weights for each feature in each factor for a given view
}
\details{
fill this
}
