% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showWeights.R
\name{showWeights}
\alias{showWeights}
\title{showWeights: get the loadings in a specific view}
\usage{
showWeights(model, view, features = "all", factors = "all", main = NULL,
  ...)
}
\arguments{
\item{model}{a fitted MOFA model}

\item{...}{further arguments that can be passed to pheatmap}

\item{viewnm}{name of view from which to get the corresponding weights}

\item{showHeatmap}{boolean, wether to plot a heatmap of the weights (default: True)}
}
\value{
a weight matrix of dimension d (feautres) x k (number of latent factors)
}
\description{
Function to extract the weights that each feautre has on a factor in the view specified and visualize it as a heatmap.
}
\details{
fill this
}
