% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showWeights.R
\name{showWeights}
\alias{showWeights}
\title{showWeights: visualise the loadings for a certain factor in a given view}
\usage{
showWeights(model, view, factor, ntop = 0, ntail = 0, manual = NULL,
  th = NULL, ntotal = 10, FeatureGroups = NULL,
  FeatureGroupsCols = NULL, showFeatureNames = F, main = NULL)
}
\arguments{
\item{model}{a fitted MOFA model}

\item{view}{name of view from which to get the corresponding weights}

\item{factor}{name of the factor}

\item{ntop}{number of highest positive weights to label}

\item{ntail}{number of lowest negative weights to label}

\item{manual}{feature names which are labeled in the plot}

\item{th}{threshold on the absolute value beyond which weigths are labeled}

\item{ntotal}{number of features with highest absolute value to label}

\item{FeatureGroups}{a named vector of same lengh as number of feautres annotating feature to groups that 
will be colored in the dotplot, names have to be in accordance with feature names}

\item{FeatureGroupsCols}{a named vector speciying for each elemnt in FeautreGroups a color to use for plotting}

\item{showFeatureNames}{boolean wether to show all FeatureNames on the x-axis (default false)}

\item{main}{plot title}
}
\description{
Function to visualize weights in a dotplot highlighting features with highest loadings
}
\details{
fill this
}
