% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetFactors.R
\name{subsetFactors}
\alias{subsetFactors}
\title{subsetFactors: subset factors}
\usage{
subsetFactors(object, factors)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{factors}{vector with the factors indices (numeric) or factor names (character) to subset.}
}
\description{
Method to subset or sort factors
}
