% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subsetViews}
\alias{subsetViews}
\title{Subset views}
\usage{
subsetViews(object, views)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{views}{character vector with the view names, numeric vector with the view indices or logical vector with the view to be kept as TRUE.}
}
\description{
Method to subset (or sort) views
}
