% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subsetViews}
\alias{subsetViews}
\title{Subset views}
\usage{
subsetViews(object, views)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{views}{character vector with the view names, numeric vector with the view indices or logical vector with the view to be kept as TRUE.}
}
\description{
Method to subset (or sort) views.
This function can remove entire views from the model. 
For example, you might want to generate the \code{\link{plotVarianceExplained}} plot excluding a particular view. \cr
This functionality is only for exploratory purposes. If some view(s) are not of interest we strongly recommend 
removing them before training the model.
}
