% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{TrainOpts}
\alias{TrainOpts}
\alias{TrainOpts,MOFAmodel-method}
\title{TrainOpts: set and retrieve training opts}
\usage{
TrainOpts(object)

\S4method{TrainOpts}{MOFAmodel}(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
list of training options
}
