% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{TrainStats}
\alias{TrainStats}
\alias{TrainStats,MOFAmodel-method}
\title{TrainStats: set and retrieve training statistics}
\usage{
TrainStats(object)

\S4method{TrainStats}{MOFAmodel}(object)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}
}
\value{
list of training statistics
}
\author{
Ricard Argelaguet
}

